Changes to taxi client protocol with a customer to taxi ping interface, tracking of taxis per customer and customers per taxi and saving/updating taxi states for a customer to facililate correspondence between a customer and nearby taxis falling within his area returned by the server. Still some gaps remaining for the ping taxi response from taxi back to the customer and ping intimation from a taxi to the remaining shortlisted taxis with the new updated shortlisted state when the taxi state transitions to PICKUP mode.

+++ taxi_test.c
@@ -11,21 +11,28 @@
 #define ID_WIDTH _STR(MAX_ID_LEN)
 
 #define TEST_FILE_NAME "locations.txt"
-
+#define MAKE_TEST_MASK(m) (1 << (m))
+#define CHECK_TEST_MASK(m, b) ( (m) & (1<<(b)) )
 static struct taxi *taxis;
 int num_taxis;
 static struct taxi_test_args
 {
+#define TEST_ADD  (0x0)
+#define TEST_DELETE (0x1)
+#define TEST_PING (0x2)
+#define TEST_SEARCH (0x3)
     char server[20];
     int port;
-    int test_delete;
+    unsigned int test_mask;
     char fname[20];
 } taxi_test_args = { .server = _TAXI_SERVER_IP, .port = _TAXI_SERVER_PORT, 
-                     .test_delete = 0, .fname = TEST_FILE_NAME ,
+                     .test_mask = MAKE_TEST_MASK(TEST_ADD) | MAKE_TEST_MASK(TEST_SEARCH),
+                     .fname = TEST_FILE_NAME ,
 };
 
 static int add_taxi_location(double latitude, double longitude, unsigned char *id, int id_len)
 {
+    int err = 0;
     taxis = realloc(taxis, sizeof(*taxis) * (num_taxis+1));
     assert(taxis);
     taxis[num_taxis].latitude = latitude;
@@ -34,17 +41,22 @@ static int add_taxi_location(double latitude, double longitude, unsigned char *i
     taxis[num_taxis].id_len = len;
     memcpy(taxis[num_taxis].id, id, len);
     num_taxis++;
-    return update_taxi_location(&taxis[num_taxis-1]);
+    if(CHECK_TEST_MASK(taxi_test_args.test_mask, TEST_ADD))
+        err = update_taxi_location(&taxis[num_taxis-1]);
+    return err;
 }
 
 static int del_taxis(void)
 {
-    for(int i = 0; i < num_taxis; ++i)
+    if(CHECK_TEST_MASK(taxi_test_args.test_mask, TEST_DELETE))
     {
-        int ret = delete_taxi(&taxis[i]);
-        assert(ret == 0);
-        printf("Taxi [%.*s] with latitude [%lg], longitude [%lg] deleted successfully\n", 
-               taxis[i].id_len, taxis[i].id, taxis[i].latitude, taxis[i].longitude);
+        for(int i = 0; i < num_taxis; ++i)
+        {
+            int ret = delete_taxi(&taxis[i]);
+            assert(ret == 0);
+            printf("Taxi [%.*s] with latitude [%lg], longitude [%lg] deleted successfully\n", 
+                   taxis[i].id_len, taxis[i].id, taxis[i].latitude, taxis[i].longitude);
+        }
     }
     free(taxis);
     num_taxis = 0;
@@ -57,8 +69,10 @@ static void display_taxis(struct taxi *taxis, int num_taxis)
     output("----Displaying [%d] taxis------\n", num_taxis);
     for(int i = 0; i < num_taxis; ++i)
     {
-        output(" Taxi [%.*s] found near [%G:%G]\n", taxis[i].id_len, taxis[i].id, 
-               taxis[i].latitude, taxis[i].longitude);
+        output(" Taxi [%.*s] found near [%G:%G] at ip [%s], port [%d]\n", 
+               taxis[i].id_len, taxis[i].id, 
+               taxis[i].latitude, taxis[i].longitude,
+               inet_ntoa(taxis[i].addr.sin_addr), ntohs(taxis[i].addr.sin_port));
     }
 }
 
@@ -76,12 +90,23 @@ static int find_taxis(struct taxi *search_taxis, int num_searches)
         if(num_taxis > 0)
         {
             display_taxis(taxis, num_taxis);
-            if(taxi_test_args.test_delete && num_taxis >= 2)
+            if(CHECK_TEST_MASK(taxi_test_args.test_mask, TEST_DELETE) 
+               && num_taxis >= 2)
             {
                 printf("Re-running the query with deletion of id [%.*s]\n",
                        taxis[0].id_len, taxis[0].id);
                 delete_taxi(&taxis[0]);
                 i--;
+            }
+            else if(CHECK_TEST_MASK(taxi_test_args.test_mask, TEST_PING))
+            {
+                printf("Testing ping command send to the taxis nearby\n");
+                struct taxi customer = {0};
+                strncat((char*)customer.id, "foobar", sizeof(customer.id)-1);
+                customer.id_len = strlen((const char*)customer.id);
+                customer.latitude = 38.4;
+                customer.longitude = -122.31;
+                ping_nearby_taxis(&customer, taxis, num_taxis);
             }
             free(taxis);
         }
@@ -132,18 +157,19 @@ static int test_taxi_scan(const char *fname)
     fclose(fptr);
     if(num_searches > 0)
     {
-        find_taxis(search_taxis, num_searches);
+        if(CHECK_TEST_MASK(taxi_test_args.test_mask, TEST_SEARCH))
+            find_taxis(search_taxis, num_searches);
         free(search_taxis);
     }
-    if(taxi_test_args.test_delete)
-        del_taxis();
+    del_taxis();
     return 0;
 }
 
 static char *prog;
 static void usage(void)
 {
-    fprintf(stderr, "%s [ -s | server ] [ -p | port ] [ -d | test deletion ] [ -h | this help ]\n",
+    fprintf(stderr, "%s [ -s | server ] [ -p | port ] [ -d | test deletion ] "
+            " [ -a | test add ] [ -i | test ping ] [ -f | test search ] [ -h | this help ]\n",
             prog);
     exit(1);
 }
@@ -151,13 +177,15 @@ static void usage(void)
 int main(int argc, char **argv)
 {
     int c;
+    unsigned int test_mask = 0;
+    int loop = 0;
     char *s;
     prog = argv[0];
     if( (s = strrchr(prog, '/') ) )
         prog = s+1;
 
     opterr = 0;
-    while ( (c = getopt(argc, argv, "s:p:dh") ) != EOF )
+    while ( (c = getopt(argc, argv, "s:p:dafihw") ) != EOF )
     {
         switch(c)
         {
@@ -171,7 +199,23 @@ int main(int argc, char **argv)
             break;
             
         case 'd':
-            taxi_test_args.test_delete = 1;
+            test_mask |= MAKE_TEST_MASK(TEST_DELETE);
+            break;
+
+        case 'a':
+            test_mask |= MAKE_TEST_MASK(TEST_ADD);
+            break;
+            
+        case 'i':
+            test_mask |= MAKE_TEST_MASK(TEST_PING);
+            break;
+
+        case 'f':
+            test_mask |= MAKE_TEST_MASK(TEST_SEARCH);
+            break;
+
+        case 'w':
+            loop = 1;
             break;
 
         case 'h':
@@ -186,7 +230,10 @@ int main(int argc, char **argv)
         taxi_test_args.fname[0] = 0;
         strncat(taxi_test_args.fname, argv[optind], sizeof(taxi_test_args.fname)-1);
     }
-
+    if(test_mask)
+    {
+        taxi_test_args.test_mask = test_mask;
+    }
     int err = taxi_client_initialize(taxi_test_args.server, taxi_test_args.port);
     if(err < 0)
     {
@@ -194,5 +241,8 @@ int main(int argc, char **argv)
         return -1;
     }
     test_taxi_scan(taxi_test_args.fname);
+    if(loop)
+        for(;;) sleep(3);
+
     return 0;
 }
